generative-ai/gemini/sample-apps/gemini-live-telephony-app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==23.0.0
python-dotenv==1.0.1
google-genai==1.28.0
twilio==9.8.6
websockets==15.0.1
numpy==2.3.4
samplerate==0.2.2
aiohttp==3.9.1
